refactor: Reorder generator store file structure
