+++ jsconfig.json
@@ -6,6 +6,6 @@
       "$lib/*": ["src/lib/*"]
     }
   },
-  "include": ["src/**/*.d.ts", "src/**/*.js", "src/**/*.svelte", "src/lib/generator/stores/IPFStokenStore.ts"],
+  "include": ["src/**/*.d.ts", "src/**/*.js", "src/**/*.svelte", "src/lib/stores/generator/IPFStokenStore.ts"],
   "extends": "./.svelte-kit/tsconfig.json"
 }
