+++ src/lib/stores/generator/interfaces.ts
